Add support for Unicode identifiers
This is a bit tricky, because for Python versions prior to 3.3 (or PEP 393)
Python may be built with strings represented in UTF-16 (narrow unicode
support) or UTF-32 (wide unicode support, PEP 261).  Since Python 3.3, all
strings should behave as though they were UTF-32, no matter the internal
representation.  But we still aim to support Python 2.7, so we do both.